actualizacion

+++ src/controllers/adminControllers/noticia.controller.ts
@@ -35,6 +35,18 @@ export class NoticiaController {
 
     @Post('delete')
     async deleteUser(@Request() req, @Body('ID') ID: number): Promise<Result> {
+        const entidad = await this.noticiaService.list({ INIT: 0, ROWS: 1, DESC: null, CESTDO: null, ID });
+        if (entidad.length === 0) {
+            return { MESSAGE: 'La noticia no existe', STATUS: false };
+        }
+
+
+        // let IMAGEN = entidad[0].IMAGEN;
+        // const deleteFile = await this.s3Service.deleteFile(IMAGEN);
+        // if (!deleteFile) {
+        //     return { MESSAGE: 'Error al eliminar la imagen', STATUS: false };
+        // }
+
         return await this.noticiaService.delete(ID, req.user.UCRCN);
     }
 
@@ -57,7 +69,7 @@ export class NoticiaController {
             }
         }),
     )
-    async addUser(@Request() req, @Body() entidad: NoticiaModel, @UploadedFiles() files): Promise<Result> {
+    async addNoticia(@Request() req, @Body() entidad: NoticiaModel, @UploadedFiles() files): Promise<Result> {
         entidad.UCRCN = req.user.UCRCN;
         try {
 
@@ -100,8 +112,63 @@ export class NoticiaController {
     }
 
     @Post('edit')
-    async editUser(@Request() req, @Body() entidad: NoticiaModel): Promise<Result> {
-        entidad.UCRCN = req.user.UCRCN;
-        return await this.noticiaService.edit(entidad);
+    @UseInterceptors(
+        FilesInterceptor('files', 20, {
+            storage: diskStorage({
+                destination: './uploads',
+                filename: function (req, file, cb) {
+                    const filename = `${Date.now()}-${file.originalname.replace(/\s/g, '')}`;
+                    return cb(null, filename);
+                }
+            }),
+            fileFilter: (req, file, cb) => {
+                if (file.mimetype.match(/\/(jpg|jpeg|png|gif)$/)) {
+                    cb(null, true);
+                } else {
+                    cb(new Error('Solo se permiten imagenes'), false);
+                }
+            }
+        }),
+    )
+    async editNoticia(@Request() req, @Body() entidad: NoticiaModel, @UploadedFiles() files?: any[]): Promise<Result> {
+        try {
+            const table: DataTable = {
+                INIT: 0,
+                ROWS: 1,
+                DESC: entidad.TITULO,
+                CESTDO: null,
+                ID: entidad.ID
+            };
+
+            const obtener = await this.noticiaService.list(table);
+            if (obtener.length > 0) {
+                return { MESSAGE: `Ya existe una noticia con el titulo ${entidad.TITULO}`, STATUS: false };
+            }
+
+            const [file1] = files;
+            
+            if (![undefined, null].includes(file1)) {
+                // const resDelete = await this.s3Service.deleteFile(entidad.IMAGEN);
+                const keysLocation: string = await this.s3Service.uploadImage(
+                    entidad,
+                    file1.filename,
+                    file1.path
+                );
+
+                entidad.IMAGEN = keysLocation;
+            }
+
+            entidad.UCRCN = req.user.UCRCN;
+            const result = await this.noticiaService.edit(entidad);
+            return result
+        }
+        catch (error) {
+            return { MESSAGE: error.message, STATUS: false };
+        }
+        finally {
+            await files.forEach(file => {
+                fs.unlinkSync(file.path);
+            });
+        }
     }
 }
